Refactor t2t_highlight.py to highlight markdown.
This file still really feels like a hack. I looked around for decent
alternatives and couldn't really find one that I liked. I considered
pygments, pygtkcodebuffer (what this module was based on), integrating
gedit, and gtkSourceView (the gtk object gedit uses).

gtkSourceView seeems like the best option, but in many ways it is
overkill. I only need syntax highlighting for one language and this
gtkCodeBuffer based module allows for the easiest and most
customization.

So, I am stuck with this. Some refactoring is certainly still needed.
There are a number of TODO notes with areas to focus on. The biggest
needed improvement is multiline highlighting.